Introduce ProcessedEventsAwareProjectorInterface
and `AbstractProcessedEventsAwareProjector` that implements the interface and
is now the base class for the 3 main projectors
