+++ Neos.EventSourcedContentRepository/Classes/Domain/Projection/ContentStream/ContentStreamProjector.php
@@ -12,8 +12,6 @@ namespace Neos\EventSourcedContentRepository\Domain\Projection\ContentStream;
  * source code.
  */
 
-use Doctrine\ORM\EntityManagerInterface;
-use Neos\Cache\Frontend\VariableFrontend;
 use Neos\EventSourcedContentRepository\Domain\Context\ContentStream\Event\ContentStreamWasCreated;
 use Neos\EventSourcedContentRepository\Domain\Context\ContentStream\Event\ContentStreamWasForked;
 use Neos\EventSourcedContentRepository\Domain\Context\ContentStream\Event\ContentStreamWasRemoved;
@@ -21,62 +19,21 @@ use Neos\EventSourcedContentRepository\Domain\Context\Workspace\Event\RootWorksp
 use Neos\EventSourcedContentRepository\Domain\Context\Workspace\Event\WorkspaceRebaseFailed;
 use Neos\EventSourcedContentRepository\Domain\Context\Workspace\Event\WorkspaceWasCreated;
 use Neos\EventSourcedContentRepository\Domain\Context\Workspace\Event\WorkspaceWasRebased;
-use Neos\EventSourcing\Event\DecoratedEvent;
-use Neos\EventSourcing\Event\DomainEvents;
-use Neos\EventSourcing\EventListener\AfterInvokeInterface;
-use Neos\EventSourcing\EventListener\AppliedEventsStorage\AppliedEventsStorageInterface;
-use Neos\EventSourcing\EventListener\AppliedEventsStorage\DoctrineAppliedEventsStorage;
-use Neos\EventSourcing\EventListener\Exception\HighestAppliedSequenceNumberCantBeReservedException;
-use Neos\EventSourcing\EventStore\EventEnvelope;
-use Neos\Flow\Annotations as Flow;
-use Neos\EventSourcing\Projection\ProjectorInterface;
+use Neos\EventSourcedContentRepository\Infrastructure\Projection\AbstractProcessedEventsAwareProjector;
 
-/**
- * Workspace Projector
- * @Flow\Scope("singleton")
- */
-class ContentStreamProjector implements ProjectorInterface, AfterInvokeInterface, AppliedEventsStorageInterface
+class ContentStreamProjector extends AbstractProcessedEventsAwareProjector
 {
     private const TABLE_NAME = 'neos_contentrepository_projection_contentstream_v1';
 
-    /**
-     * @var \Doctrine\DBAL\Connection
-     */
-    private $dbal;
-
-    /**
-     * @Flow\Inject
-     * @var VariableFrontend
-     */
-    protected $processedEventsCache;
-
-    /**
-     * @var bool
-     */
-    protected $assumeProjectorRunsSynchronously = false;
-
-    public function injectEntityManager(EntityManagerInterface $entityManager): void
+    public function reset(): void
     {
-        $this->dbal = $entityManager->getConnection();
-        $this->doctrineAppliedEventsStorage = new DoctrineAppliedEventsStorage($this->dbal, ContentStreamProjector::class);
-    }
-
-    /**
-     * @internal
-     */
-    public function assumeProjectorRunsSynchronously()
-    {
-        $this->assumeProjectorRunsSynchronously = true;
-    }
-
-    public function isEmpty(): bool
-    {
-        return false;
+        parent::reset();
+        $this->getDatabaseConnection()->exec('TRUNCATE ' . self::TABLE_NAME);
     }
 
     public function whenContentStreamWasCreated(ContentStreamWasCreated $event)
     {
-        $this->dbal->insert(self::TABLE_NAME, [
+        $this->getDatabaseConnection()->insert(self::TABLE_NAME, [
             'contentStreamIdentifier' => $event->getContentStreamIdentifier(),
             'state' => ContentStreamFinder::STATE_CREATED,
         ]);
@@ -85,7 +42,7 @@ class ContentStreamProjector implements ProjectorInterface, AfterInvokeInterface
     public function whenRootWorkspaceWasCreated(RootWorkspaceWasCreated $event)
     {
         // the content stream is in use now
-        $this->dbal->update(self::TABLE_NAME, [
+        $this->getDatabaseConnection()->update(self::TABLE_NAME, [
             'state' => ContentStreamFinder::STATE_IN_USE_BY_WORKSPACE,
         ], [
             'contentStreamIdentifier' => $event->getCurrentContentStreamIdentifier()
@@ -95,7 +52,7 @@ class ContentStreamProjector implements ProjectorInterface, AfterInvokeInterface
     public function whenWorkspaceWasCreated(WorkspaceWasCreated $event)
     {
         // the content stream is in use now
-        $this->dbal->update(self::TABLE_NAME, [
+        $this->getDatabaseConnection()->update(self::TABLE_NAME, [
             'state' => ContentStreamFinder::STATE_IN_USE_BY_WORKSPACE,
         ], [
             'contentStreamIdentifier' => $event->getCurrentContentStreamIdentifier()
@@ -104,7 +61,7 @@ class ContentStreamProjector implements ProjectorInterface, AfterInvokeInterface
 
     public function whenContentStreamWasForked(ContentStreamWasForked $event)
     {
-        $this->dbal->insert(self::TABLE_NAME, [
+        $this->getDatabaseConnection()->insert(self::TABLE_NAME, [
             'contentStreamIdentifier' => $event->getContentStreamIdentifier(),
             'sourceContentStreamIdentifier' => $event->getSourceContentStreamIdentifier(),
             'state' => ContentStreamFinder::STATE_REBASING,
@@ -115,14 +72,14 @@ class ContentStreamProjector implements ProjectorInterface, AfterInvokeInterface
     public function whenWorkspaceWasRebased(WorkspaceWasRebased $event)
     {
         // the new content stream is in use now
-        $this->dbal->update(self::TABLE_NAME, [
+        $this->getDatabaseConnection()->update(self::TABLE_NAME, [
             'state' => ContentStreamFinder::STATE_IN_USE_BY_WORKSPACE,
         ], [
             'contentStreamIdentifier' => $event->getCurrentContentStreamIdentifier()
         ]);
 
         // the previous content stream is no longer in use
-        $this->dbal->update(self::TABLE_NAME, [
+        $this->getDatabaseConnection()->update(self::TABLE_NAME, [
             'state' => ContentStreamFinder::STATE_NO_LONGER_IN_USE,
         ], [
             'contentStreamIdentifier' => $event->getPreviousContentStreamIdentifier()
@@ -131,7 +88,7 @@ class ContentStreamProjector implements ProjectorInterface, AfterInvokeInterface
 
     public function whenWorkspaceRebaseFailed(WorkspaceRebaseFailed $event)
     {
-        $this->dbal->update(self::TABLE_NAME, [
+        $this->getDatabaseConnection()->update(self::TABLE_NAME, [
             'state' => ContentStreamFinder::STATE_REBASE_ERROR,
         ], [
             'contentStreamIdentifier' => $event->getTargetContentStreamIdentifier()
@@ -140,85 +97,10 @@ class ContentStreamProjector implements ProjectorInterface, AfterInvokeInterface
 
     public function whenContentStreamWasRemoved(ContentStreamWasRemoved $event)
     {
-        $this->dbal->update(self::TABLE_NAME, [
+        $this->getDatabaseConnection()->update(self::TABLE_NAME, [
             'removed' => true
         ], [
             'contentStreamIdentifier' => $event->getContentStreamIdentifier()
         ]);
-    }
-
-    public function reset(): void
-    {
-        $this->dbal->transactional(function () {
-            $this->dbal->exec('TRUNCATE ' . self::TABLE_NAME);
-        });
-    }
-
-    /**
-     * Called after a listener method is invoked
-     *
-     * @param EventEnvelope $eventEnvelope
-     * @return void
-     */
-    public function afterInvoke(EventEnvelope $eventEnvelope): void
-    {
-        if ($this->assumeProjectorRunsSynchronously === true) {
-            // if we run synchronously during an import, we don't need the processed events cache
-            return;
-        }
-        $this->processedEventIdentifiers[] = $eventEnvelope->getRawEvent()->getIdentifier();
-    }
-
-    public function hasProcessed(DomainEvents $events): bool
-    {
-        if ($this->assumeProjectorRunsSynchronously === true) {
-            // if we run synchronously during an import, we *know* that events have been processed already.
-            return true;
-        }
-        foreach ($events as $event) {
-            if (!$event instanceof DecoratedEvent) {
-                throw new \RuntimeException(sprintf('The CommandResult contains an event "%s" that is no DecoratedEvent', get_class($event)), 1550314769);
-            }
-            if (!$this->processedEventsCache->has(md5($event->getIdentifier()))) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    /**
-     * @var string[]
-     */
-    private $processedEventIdentifiers = [];
-    /**
-     * @var DoctrineAppliedEventsStorage
-     */
-    private $doctrineAppliedEventsStorage;
-
-
-    /**
-     * @inheritDoc
-     */
-    public function reserveHighestAppliedEventSequenceNumber(): int
-    {
-        return $this->doctrineAppliedEventsStorage->reserveHighestAppliedEventSequenceNumber();
-    }
-    /**
-     * @inheritDoc
-     */
-    public function releaseHighestAppliedSequenceNumber(): void
-    {
-        $this->doctrineAppliedEventsStorage->releaseHighestAppliedSequenceNumber();
-        foreach ($this->processedEventIdentifiers as $eventIdentifier) {
-            $this->processedEventsCache->set(md5($eventIdentifier), true);
-        }
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function saveHighestAppliedSequenceNumber(int $sequenceNumber): void
-    {
-        $this->doctrineAppliedEventsStorage->saveHighestAppliedSequenceNumber($sequenceNumber);
     }
 }
