+++ Neos.EventSourcedContentRepository/Classes/Infrastructure/Projection/ProcessedEventsAwareProjectorInterface.php
@@ -1,0 +1,16 @@
+<?php
+declare(strict_types=1);
+namespace Neos\EventSourcedContentRepository\Infrastructure\Projection;
+
+use Neos\EventSourcing\Event\DomainEvents;
+use Neos\EventSourcing\EventListener\AfterInvokeInterface;
+use Neos\EventSourcing\EventListener\AppliedEventsStorage\AppliedEventsStorageInterface;
+use Neos\EventSourcing\Projection\ProjectorInterface;
+
+interface ProcessedEventsAwareProjectorInterface extends ProjectorInterface, AfterInvokeInterface, AppliedEventsStorageInterface
+{
+
+    public function assumeProjectorRunsSynchronously(): void;
+
+    public function hasProcessed(DomainEvents $events): bool;
+}
